Created a framework for a class that automatically averages arrays for SRS hub purposes.

+++ TeamCode/src/main/java/org/nknsd/teamcode/components/handlers/srs/SRSDataAverager.java
@@ -10,7 +10,7 @@ import org.nknsd.teamcode.frameworks.NKNComponent;
 import java.util.concurrent.Callable;
 
 public class SRSDataAverager implements NKNComponent {
-    public int[][] getAveragedData() {
+    public short[][] getAveragedData() {
         return null;
     }
